requirements.txt
streamlit==1.28.1
pdfplumber==0.10.3
pypdf==3.17.1
tiktoken==0.5.2
python-dotenv==1.0.0
pytest==7.4.3
openai==1.51.0
chromadb==0.4.17
